refactor(map): 상수 및 훅 파일명 변경
